Add custom mkYarnPackage facade without link farming
This commit adds a `mkYarnPackage` alternative which does not perform
link farming and any additional indirection that `yarn2nix` normally
does, which is extremely brittle on large projects (especially those
with Git dependencies containing Git submodules). This API closely
mimics the actual `mkYarnPackage` function and saves it in a file called
`tenx.nix` to differentiate it from the rest of the project.